is rdy

+++ src/Book.java
@@ -1,9 +1,9 @@
 public class Book {
     private String title;
     private String author;
-    private int price;
+    private double price;
 
-    public Book(String title, String author, int price) {
+    public Book(String title, String author, double price) {
         this.title = title;
         this.author = author;
         this.price = price;
@@ -25,7 +25,7 @@ public class Book {
         this.author = author;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
